Move files to correct package
