+++ page-controller/src/main/java/com/iluwatar/page/controller/SignupModel.java
@@ -24,28 +24,19 @@
  */
 package com.iluwatar.page.controller;
 
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.stereotype.Component;
 
 /**
- * User Controller.
+ * ignup model.
  */
-@Slf4j
-@Controller
-public class UserController {
-  private final UserView view = new UserView();
+@Component
+@Data
+@NoArgsConstructor
+public class SignupModel {
+  private String name;
+  private String email;
+  private String password;
 
-  public UserController() {}
-
-  /**
-   * Handle http GET request and access view and model.
-   */
-  @GetMapping("/user")
-  public String getUserPath(SignupModel form, Model model) {
-    model.addAttribute("name", form.getName());
-    model.addAttribute("email", form.getEmail());
-    return view.display(form);
-  }
 }
